Restructure folders and Some edits
