--- README.md
@@ -1,2 +1,0 @@
-# Iot_system_DXC
-Comprehensive IoT Solution: The Smart IoT Monitoring System collects, processes, and analyzes real-time sensor data, serving industries like smart homes, agriculture, and environmental monitoring.
