+++ backend/src/main/java/com/DXC/iotbackend/IotSystemDxcApplication.java
@@ -1,0 +1,13 @@
+package com.DXC.iotbackend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IotSystemDxcApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(IotSystemDxcApplication.class, args);
+    }
+
+}
