+++ backend/src/main/java/com/DXC/iotbackend/JwtFilter.java
@@ -1,0 +1,84 @@
+package com.DXC.iotbackend;
+
+import com.DXC.iotbackend.util.JwtUtil;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Component
+public class JwtFilter extends OncePerRequestFilter {
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request,
+                                    HttpServletResponse response,
+                                    FilterChain filterChain)
+            throws ServletException, IOException {
+        System.out.println("🔎 Incoming Headers:");
+//        Collections.list(request.getHeaderNames()).forEach(h ->
+//                System.out.println("→ " + h + ": " + request.getHeader(h))
+//        );
+
+
+        String authHeader = request.getHeader("Authorization");
+        System.out.println(" Authorization header: " + authHeader);
+        System.out.println(" JwtFilter triggered for URI: " + request.getRequestURI());
+
+        if (authHeader != null && authHeader.startsWith("Bearer ")) {
+            String token = authHeader.substring(7);
+            System.out.println("🔑 JWT token: " + token);
+
+            try {
+                String username = jwtUtil.extractUsername(token);
+                String Role = jwtUtil.extractRole(token);
+                System.out.println(" Username from token: " + username);
+                System.out.println(" Roles from token: " + Role);
+
+                if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+                    List<GrantedAuthority> authorities = List.of(new SimpleGrantedAuthority("ROLE_" + Role));
+
+
+                    UserDetailsImpl userDetails = new UserDetailsImpl(username, Role);
+
+                    UsernamePasswordAuthenticationToken authToken =
+                            new UsernamePasswordAuthenticationToken(userDetails, null, authorities);
+
+                    authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                    SecurityContextHolder.getContext().setAuthentication(authToken);
+
+                    System.out.println(" Authenticated user: " + username);
+                }
+
+            } catch (ExpiredJwtException e) {
+                response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+                response.getWriter().write("Token expired");
+                System.out.println(" Token expired");
+                return;
+            } catch (Exception e) {
+                response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+                response.getWriter().write("Invalid token");
+                System.out.println(" Invalid token: " + e.getMessage());
+                return;
+            }
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}
