+++ backend/src/main/java/com/DXC/iotbackend/controller/AppController.java
@@ -1,0 +1,63 @@
+package com.DXC.iotbackend.controller;
+
+import com.DXC.iotbackend.UserDto;
+import com.DXC.iotbackend.model.UserEntity;
+import com.DXC.iotbackend.payload.LoginRequest;
+import com.DXC.iotbackend.repository.UserRepository;
+import com.DXC.iotbackend.service.AuthService;
+import com.DXC.iotbackend.util.JwtUtil;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Map;
+
+@RestController
+public class AppController {
+    @Autowired
+    private AuthService authService;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @PostMapping("/register")
+    public void register(@RequestBody @Valid UserDto user) {
+        System.out.println(" Register endpoint hit!");
+        authService.registerUser(user);
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<?> login(@RequestBody LoginRequest loginRequest) {
+        UserEntity user = userRepository.findByEmail(loginRequest.getEmail())
+                .orElseThrow(() -> new RuntimeException("User not found"));
+
+        if (!passwordEncoder.matches(loginRequest.getPassword(), user.getPassword())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid credentials");
+        }
+
+        String token = jwtUtil.generateToken(user.getUsername(), user.getRole());
+        return ResponseEntity.ok(Map.of("token", token));
+    }
+
+
+    @GetMapping("/api/user")
+    public ResponseEntity<?> getUser(Authentication auth) {
+        return ResponseEntity.ok("Welcome, " + auth.getName());
+    }
+
+
+
+}
