+++ backend/src/main/java/com/DXC/iotbackend/model/UserEntity.java
@@ -1,0 +1,67 @@
+package com.DXC.iotbackend.model;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "userssss")
+public class UserEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String username;
+    private String password;
+    private String email;
+    private String role; // "USER" or "ADMIN"
+
+
+    private String firstName;
+    private String lastName;
+
+    // Constructors
+    public UserEntity() {}
+
+    public UserEntity(String username, String password, String email,String role, String firstName, String lastName) {
+        this.username = username;
+        this.password = password;
+        this.email = email;
+        this.role=role;
+        this.firstName = firstName;
+        this.lastName = lastName;
+
+    }
+
+    // Getters & Setters
+    public Long getId() { return id; }
+
+    public String getRole() { return role; }
+    public void setRole(String role) { this.role = role; }
+
+
+    public String getUsername() { return username; }
+    public void setUsername(String username) { this.username = username; }
+
+    public String getPassword() { return password; }
+    public void setPassword(String password) { this.password = password; }
+
+    public String getEmail() { return email; }
+    public void setEmail(String email) { this.email = email; }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+}
