+++ backend/src/main/java/com/DXC/iotbackend/payload/LoginRequest.java
@@ -1,0 +1,23 @@
+package com.DXC.iotbackend.payload;
+
+public class LoginRequest {
+    private String email;
+    private String password;
+
+    // Getters and setters
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
