--- demo/src/main/java/com/example/demo/IotSystemDxcApplication.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IotSystemDxcApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(IotSystemDxcApplication.class, args);
-    }
-
-}
