--- demo/src/main/java/com/example/demo/LoginRequest.java
@@ -1,23 +1,0 @@
-package com.example.demo;
-
-public class LoginRequest {
-    private String email;
-    private String password;
-
-    // Getters and setters
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
