--- demo/src/main/java/com/example/demo/RateLimitingFilter.java
@@ -1,37 +1,0 @@
-package com.example.demo;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-
-//@Order(0)
-@Component
-public class RateLimitingFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private RedisRateLimiter rateLimiter;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                    HttpServletResponse response,
-                                    FilterChain filterChain) throws ServletException, IOException {
-
-        String ipAddress = request.getRemoteAddr();
-
-        if (!rateLimiter.isAllowed(ipAddress)) {
-            response.setStatus(429);
-            response.getWriter().write("Rate limit exceeded. Try again later.");
-            return;
-        }
-
-        filterChain.doFilter(request, response);
-    }
-}
