--- demo/src/main/resources/application.properties
@@ -1,29 +1,0 @@
-#spring.application.name=IOT_System_DXC
-#spring.datasource.url=jdbc:h2:mem:testdb
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=sa
-#spring.datasource.password=
-#spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-#spring.h2.console.enabled=true
-#spring.jpa.hibernate.ddl-auto=update
-
-
-# MySQL DB Connection
-spring.datasource.url=jdbc:mysql://localhost:3306/dxc?useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=
-
-# JPA Settings
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
-
-# Disable H2 Console (optional)
-spring.h2.console.enabled=false
-
-
-spring.devtools.restart.enabled=false
-jwt.secret=3cfa76ef14937c1c0ea519f8fc057a80fcd04a7420f8e8bcd0a7567c272e007b
-
-spring.data.redis.host=localhost
-spring.data.redis.port=6379
